+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SDTable` now emits `toggleExpanders` when a row is expanded, sending the expanded row data as payload.
 
 ## 3.0.0-beta.1 - 2026-01-27
 ### Added